requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
attrs==25.3.0
beanie==1.30.0
blinker==1.9.0
Cerberus==1.3.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
crispy-bootstrap5==2025.4
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
Django==4.2.5
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
environs==14.1.1
execnet==2.1.1
fastapi==0.115.11
fastapi_cors==0.0.6
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.1
fonttools==4.57.0
frozenlist==1.7.0
fsspec==2025.3.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
lazy-model==0.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.77.0
orjson==3.11.0
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250324
pgvector==0.4.0
pillow==11.1.0
pipenv==2025.0.2
platformdirs==4.3.8
pluggy==1.5.0
portalocker==2.10.1
priority==2.0.0
propcache==0.3.2
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
PyJWT==2.9.0
pymongo==4.15.4
PyMuPDF==1.25.4
pyparsing==3.2.3
pytest==8.3.5
pytest-shutil==1.8.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
Quart==0.20.0
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.4.4
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
termcolor==2.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.31.2
waitress==3.0.2
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
